bye backgroundImpl
